+++ include/geometry/geometry.hpp
@@ -7,7 +7,7 @@ class Geometry
 {
 public:
 
-    virtual float implicit(Geometric x) const = 0;
+    virtual double implicit(Geometric x) const = 0;
 
     virtual bool intersect_with_ray(const Ray& r, Intersection& intersection) const = 0;
