+++ include/geometry/triangle.hpp
@@ -18,7 +18,7 @@ public:
     Geometric normal() const;
     Plane plane() const ;
     
-    float implicit(Geometric x) const;
+    double implicit(Geometric x) const;
 
     bool intersect_with_ray(const Ray& r, Intersection& intersection) const;
 };
