+++ include/geometry/triangle_mesh.hpp
@@ -16,7 +16,7 @@ public:
 
     TriangleMesh(std::vector<Triangle> triangles);
     
-    float implicit(Geometric x) const;
+    double implicit(Geometric x) const;
 
     bool intersect_with_ray(const Ray& r, Intersection& intersection) const;
 };
