+++ include/tone_mapping.hpp
@@ -10,7 +10,7 @@
 class ToneMapping
 {
 protected:
-    float _LMax; ///< Maximum luminance value for tone mapping
+    double _LMax; ///< Maximum luminance value for tone mapping
 
 public:
     /**
@@ -18,14 +18,14 @@ public:
      * @param LMax Maximum luminance value.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    ToneMapping(float LMax);
+    ToneMapping(double LMax);
 
     /**
      * @brief Evaluate the tone mapping for a given luminance value.
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    virtual float evaluate(float l_in) const = 0;
+    virtual double evaluate(double l_in) const = 0;
 
     /**
      * @brief Retrieves the maximum luminance value.
@@ -36,7 +36,7 @@ public:
      * 
      * @return The maximum luminance value.
      */
-    float max_luminance() const;
+    double max_luminance() const;
 
 };
 
@@ -57,14 +57,14 @@ public:
      * @param LMax Maximum luminance value.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    Clamping(float LMax);
+    Clamping(double LMax);
 
     /**
      * @brief Evaluate the clamping tone mapping for a given luminance value.
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 
 };
 
@@ -85,7 +85,7 @@ public:
      * @param LMax Maximum luminance value.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    Equalization(float LMax);
+    Equalization(double LMax);
 
     /**
      * @brief Evaluate the equalization tone mapping for a given
@@ -93,7 +93,7 @@ public:
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
 
 
@@ -109,7 +109,7 @@ public:
 class EqualizationClamping : public ToneMapping
 {
 private:
-    float _V; ///< Clamping value for tone mapping
+    double _V; ///< Clamping value for tone mapping
 
 public:
     /**
@@ -119,7 +119,7 @@ public:
      * @throws std::invalid_argument if V is greater than LMax or less than 0.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    EqualizationClamping(float V, float LMax);
+    EqualizationClamping(double V, double LMax);
 
     /**
      * @brief Evaluate the equalization and the clamping tone mapping
@@ -127,7 +127,7 @@ public:
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
 
 
@@ -142,7 +142,7 @@ public:
 class Gamma : public ToneMapping
 {
 private:
-    float _gamma;   ///< Gamma value for tone mapping
+    double _gamma;   ///< Gamma value for tone mapping
     
 public:
     /**
@@ -152,14 +152,14 @@ public:
      * @throws std::invalid_argument if gamma is less than or equal to 0.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    Gamma(float gamma, float LMax);
+    Gamma(double gamma, double LMax);
 
     /**
      * @brief Evaluate the gamma tone mapping for a given luminance value.
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
 
 
@@ -174,8 +174,8 @@ public:
 class GammaClamping : public ToneMapping
 {
 private:
-    float _gamma;   ///< Gamma value for tone mapping
-    float _V;       ///< Clamping value for tone mapping
+    double _gamma;   ///< Gamma value for tone mapping
+    double _V;       ///< Clamping value for tone mapping
     
 public:
     /**
@@ -187,7 +187,7 @@ public:
      * @throws std::invalid_argument if V is greater than LMax or less than 0.
      * @throws std::invalid_argument if LMax is less than or equal to 0.
      */
-    GammaClamping(float gamma, float V, float LMax);
+    GammaClamping(double gamma, double V, double LMax);
 
     /**
      * @brief Evaluate the gamma and the clamping tone mapping for a given
@@ -195,7 +195,7 @@ public:
      * @param l_in Input luminance value.
      * @return Mapped luminance value.
      */
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
 
 
@@ -203,19 +203,19 @@ public:
 class Drago : public ToneMapping
 {
 public:
-    Drago(float LMax);
+    Drago(double LMax);
 
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
 
 
 class Logarithmic : public ToneMapping
 {
 private:
-    float _alpha;
+    double _alpha;
     
 public:
-    Logarithmic(float LMax, float alpha);
+    Logarithmic(double LMax, double alpha);
 
-    float evaluate(float l_in) const;
+    double evaluate(double l_in) const;
 };
